resolve comment
